Fix Windows build/package portability for fork runners
The previous run (32193779231) failed at:
  - Windows / windows_portable: Build step (XCOPY of C:\musescore_dependencies
    failed on a fresh windows-2025 image because the folder does not exist
    on forks / cold caches)
  - Windows / windows_x64: Package step (hardcoded uuidgen.exe path and
    reliance on %WIX% env var mean the step could not find WiX 3 tools)

Fixes:
  - build.bat: only copy C:\musescore_dependencies when it exists, log a
    warning otherwise instead of failing the whole build
  - package.bat: resolve uuidgen.exe and WiX dynamically (try where,
    fall back to powershell for GUID; search C:\Program Files\WiX* and
    C:\wix* for WiX bin). Also resolve PortableApps launcher/installer
    paths dynamically for the portable job
  - scoreorder.cpp: cast part->nstaves() to int to silence MSVC narrowing
    warning that will become an error under /WX

+++ src/engraving/dom/scoreorder.cpp
@@ -408,7 +408,7 @@ void ScoreOrder::setBracketsAndBarlines(Score* score)
 
         staff_idx_t staffIdx { 0 };
         const bool isMultiStaffInstrument = instrTemplate->staffCount > 1;
-        const int partStaffCount = part->nstaves();
+        const int partStaffCount = static_cast<int>(part->nstaves());
 
         const int templateBracketSpan = std::accumulate(instrTemplate->bracketSpan, instrTemplate->bracketSpan + MAX_STAVES, 0);
         const bool bracketSpansAllStaves = templateBracketSpan == (int)instrTemplate->staffCount;
